backend/requirements.txt
fastapi
uvicorn
python-multipart
psycopg2-binary
python-dotenv
aiofiles
sqlalchemy
asyncpg
azure-storage-blob
requests
